move metrics to components folder
