+++ packages/esm-patient-queues-app/src/components/patient-queue-metrics/clinic-metrics.scss
