+++ packages/esm-patient-queues-app/src/components/patient-queue-metrics/metrics-card.scss
