+++ packages/esm-patient-queues-app/src/components/patient-queue-metrics/metrics-header.scss
