+++ packages/esm-patient-queues-app/src/queue-reception-home.component.tsx
@@ -1,14 +1,14 @@
 import React from 'react';
 import PatientQueueHeader from './components/patient-queue-header/patient-queue-header.component';
 import ActiveVisitsReceptionTable from './active-visits/active-visits-patients-reception/active-visits-reception-table.component';
-import MetricsCard from './patient-queue-metrics/metrics-card.component';
+import MetricsCard from './components/patient-queue-metrics/metrics-card.component';
 import { useTranslation } from 'react-i18next';
 import dayjs from 'dayjs';
 import styles from './patient-queue-metrics/clinic-metrics.scss';
 
 import { useParentLocation } from './active-visits/patient-queues.resource';
 import { usePatientQueuesList } from './active-visits/active-visits-patients-reception/active-visits-reception.resource';
-import {  useServicePointCount } from './patient-queue-metrics/clinic-metrics.resource';
+import {  useServicePointCount } from './components/patient-queue-metrics/clinic-metrics.resource';
 import { useSession } from '@openmrs/esm-framework';
 
 const ReceptionHome: React.FC = () => {
